+++ modules/log/test/index.js
@@ -1,3 +1,3 @@
-import './utils/get-hi-res-timestamp.spec';
+import './lib/probe.spec';
 import './lib/normalize-arguments.spec';
 import './lib/log.spec';